requirements.txt
streamlit
google-generativeai
python-dotenv
sqlalchemy
psycopg2-binary
redshift-connector